Refine embedding naming and move dependency to 3rd_party (#1125)
