+++ .github/CODEOWNERS
@@ -3,7 +3,7 @@
 
 * liang1.lv@intel.com feng.tian@intel.com suyue.chen@intel.com
 /.github/ suyue.chen@intel.com ze.pan@intel.com
-/comps/3rd_parties/ liang1.lv@intel.com letong.han@intel.com xinyao.wang@intel.com
+/comps/third_parties/ liang1.lv@intel.com letong.han@intel.com xinyao.wang@intel.com
 /comps/agent/ kaokao.lv@intel.com minmin.hou@intel.com
 /comps/animation/ qing.yao@intel.com chun.tao@intel.com
 /comps/asr/ sihan.chen@intel.com liang1.lv@intel.com
