--- .github/workflows/docker/compose/3rd_parties-compose.yaml
@@ -1,17 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-services:
-  nginx:
-    build:
-      dockerfile: comps/3rd_parties/nginx/src/Dockerfile
-    image: ${REGISTRY:-opea}/nginx:${TAG:-latest}
-
-  wav2lip:
-    build:
-      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile
-    image: ${REGISTRY:-opea}/wav2lip:${TAG:-latest}
-  wav2lip-gaudi:
-    build:
-      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile.intel_hpu
-    image: ${REGISTRY:-opea}/wav2lip-gaudi:${TAG:-latest}
