+++ .github/workflows/docker/compose/animation-compose.yaml
@@ -9,9 +9,9 @@ services:
     image: ${REGISTRY:-opea}/animation:${TAG:-latest}
   wav2lip:
     build:
-      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile
+      dockerfile: comps/third_parties/wav2lip/src/Dockerfile
     image: ${REGISTRY:-opea}/wav2lip:${TAG:-latest}
   wav2lip-gaudi:
     build:
-      dockerfile: comps/3rd_parties/wav2lip/src/Dockerfile.intel_hpu
+      dockerfile: comps/third_parties/wav2lip/src/Dockerfile.intel_hpu
     image: ${REGISTRY:-opea}/wav2lip-gaudi:${TAG:-latest}
