--- comps/embeddings/src/integrations/dependency/bridgetower/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (C) 2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
-
-from .bridgetower_embedding import BridgeTowerEmbedding
-from .bridgetower_custom import BridgeTowerTextFeatureExtractor, BridgeTowerForITC
