+++ comps/embeddings/src/integrations/multimodal_bridgetower.py
