+++ comps/third_parties/bridgetower/deployment/docker_compose/compose_intel_cpu.yaml
