+++ comps/third_parties/bridgetower/src/Dockerfile
@@ -14,7 +14,7 @@ ENV PYTHONPATH=/home/user:/usr/lib/habanalabs/:/optimum-habana
 COPY --chown=user comps /home/user/comps
 
 RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r /home/user/comps/embeddings/src/integrations/dependency/bridgetower/requirements.txt
+    pip install --no-cache-dir -r /home/user/comps/third_parties/bridgetower/src/requirements.txt
     
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
@@ -22,6 +22,6 @@ ENV PYTHONPATH=$PYTHONPATH:/home/user
 ARG EMBEDDER_PORT=8080
 ENV PORT=$EMBEDDER_PORT
 
-WORKDIR /home/user/comps/embeddings/src/integrations/dependency/bridgetower/
+WORKDIR /home/user/comps/third_parties/bridgetower/src/
 
 ENTRYPOINT ["python", "bridgetower_server.py", "--device", "cpu"]
