+++ comps/third_parties/bridgetower/src/Dockerfile.intel_hpu
@@ -18,7 +18,7 @@ COPY --chown=user comps /home/user/comps
 
 # Install requirements and optimum habana
 RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r /home/user/comps/embeddings/src/integrations/dependency/bridgetower/requirements.txt && \
+    pip install --no-cache-dir -r /home/user/comps/third_parties/bridgetower/src/requirements.txt && \
     pip install --no-cache-dir optimum[habana]
 
 ENV PYTHONPATH=$PYTHONPATH:/home/user
@@ -26,6 +26,6 @@ ENV PYTHONPATH=$PYTHONPATH:/home/user
 ARG EMBEDDER_PORT=8080
 ENV PORT=$EMBEDDER_PORT
 
-WORKDIR /home/user/comps/embeddings/src/integrations/dependency/bridgetower/
+WORKDIR /home/user/comps/third_parties/bridgetower/src/
 ENTRYPOINT ["python", "bridgetower_server.py", "--device", "hpu"]
