+++ comps/third_parties/bridgetower/src/bridgetower_server.py
@@ -17,7 +17,7 @@ from fastapi import BackgroundTasks, FastAPI, Request
 from fastapi.responses import JSONResponse, Response
 from utils import build_logger
 
-from comps.embeddings.src.integrations.dependency.bridgetower import BridgeTowerEmbedding
+from comps.third_parties.bridgetower.src.bridgetower_embedding import BridgeTowerEmbedding
 
 worker_id = str(uuid.uuid4())[:6]
 print(f"worker_id: {worker_id}")
