+++ comps/third_parties/clip/deployment/docker_compose/compose_intel_cpu.yaml
