+++ comps/third_parties/clip/src/README.md
@@ -22,14 +22,14 @@ Users are albe to configure and build embedding-related services according to th
 
 ```bash
 cd GenAIComps/
-docker build -t opea/embedding-multimodal-clip:latest --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/embeddings/src/integrations/dependency/clip/Dockerfile .
+docker build -t opea/embedding-multimodal-clip:latest --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -f comps/third_parties/clip/src/Dockerfile .
 ```
 
 ### 1.2 Run Docker with Docker Compose
 
 ```bash
-cd comps/embeddings/src/integrations/dependency/clip/
-docker compose -f compose.yaml up -d
+cd comps/third_parties/clip/deployment/docker_compose/
+docker compose -f compose_intel_cpu.yaml up -d
 ```
 
 ## 🚀2. Consume Embedding Service
