+++ comps/third_parties/clip/src/clip_embedding.py
