+++ comps/third_parties/clip/src/clip_server.py
