+++ comps/third_parties/clip/src/requirements.txt
