+++ comps/third_parties/nginx/deployment/docker_compose/docker-compose.yaml
