+++ comps/third_parties/nginx/src/Dockerfile
@@ -6,7 +6,7 @@ FROM nginx:alpine
 
 RUN apk add --no-cache gettext
 
-COPY comps/3rd_parties/nginx/src/nginx.conf.template /etc/nginx/nginx.conf.template
+COPY comps/third_parties/nginx/src/nginx.conf.template /etc/nginx/nginx.conf.template
 
 ENV FRONTEND_SERVICE_IP=localhost
 ENV FRONTEND_SERVICE_PORT=5173
@@ -16,7 +16,7 @@ ENV BACKEND_SERVICE_PORT=8888
 ENV DATAPREP_SERVICE_IP=localhost
 ENV DATAPREP_SERVICE_PORT=6007
 
-COPY comps/3rd_parties/nginx/deployment/docker_compose/start-nginx.sh /usr/local/bin/start-nginx.sh
+COPY comps/third_parties/nginx/deployment/docker_compose/start-nginx.sh /usr/local/bin/start-nginx.sh
 RUN chmod +x /usr/local/bin/start-nginx.sh
 
 CMD ["/usr/local/bin/start-nginx.sh"]
