+++ comps/third_parties/tgi/deployment/docker_compose/launch_tgi_service.sh
