+++ comps/third_parties/wav2lip/src/Dockerfile.intel_hpu
@@ -19,7 +19,7 @@ RUN rm -rf /var/lib/apt/lists/*
 # Install GenAIComps
 RUN mkdir -p /home/user/comps
 COPY comps /home/user/comps
-COPY comps/3rd_parties/wav2lip/src/entrypoint.sh /usr/local/bin/entrypoint.sh
+COPY comps/third_parties/wav2lip/src/entrypoint.sh /usr/local/bin/entrypoint.sh
 
 # Install ffmpeg with x264 software codec
 RUN git clone https://github.com/FFmpeg/FFmpeg.git /home/user/comps/animation/FFmpeg
