+++ comps/third_parties/wav2lip/src/entrypoint.sh
@@ -23,7 +23,7 @@ export PT_HPU_LAZY_MODE=0
 export PT_HPU_ENABLE_REFINE_DYNAMIC_SHAPES=1
 
 # Wav2Lip, GFPGAN
-cd /home/user/comps/3rd_parties/wav2lip/src || exit
+cd /home/user/comps/third_parties/wav2lip/src || exit
 python3 wav2lip_server.py \
 --device $DEVICE \
 --port $((WAV2LIP_PORT)) \
