+++ tests/3rd_parties/test_3rd_parties_nginx.sh
@@ -10,7 +10,7 @@ ip_address=$(hostname -I | awk '{print $1}')
 
 function build_docker_images() {
     cd $WORKPATH
-    docker build --no-cache --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -t opea/nginx:comps -f comps/3rd_parties/nginx/src/Dockerfile .
+    docker build --no-cache --build-arg https_proxy=$https_proxy --build-arg http_proxy=$http_proxy -t opea/nginx:comps -f comps/third_parties/nginx/src/Dockerfile .
     if [ $? -ne 0 ]; then
         echo "opea/nginx built fail"
         exit 1
