+++ tests/animation/test_animation_wav2lip.sh
@@ -10,7 +10,7 @@ ip_address=$(hostname -I | awk '{print $1}')
 function build_docker_images() {
     cd $WORKPATH
     echo $(pwd)
-    docker build -t opea/wav2lip:comps -f comps/3rd_parties/wav2lip/src/Dockerfile .
+    docker build -t opea/wav2lip:comps -f comps/third_parties/wav2lip/src/Dockerfile .
     if [ $? -ne 0 ]; then
         echo "opea/wav2lip built fail"
         exit 1
